Issue #293 and #352
Issue #293

- Added archive for m_current_attribs in chardata_serializers.cpp
- Removed Character.m_current_attribs because we should be using Character.m_char_data.m_current_attribs now
- Adjusted code that used Character.m_current_attribs to used Character.m_char_data.m_current_attribs

Suggestion: Make m_max_attribs as part of CharacterData as well

Issue #352

- Commented out currently unused variables in InputState.cpp
- Removed useless casting of commands.size() to uint32_t in NetCommandManager.cpp
- Changed Team.m_team_leader_idx from int to uint32_t to remove compare warnings (also to make the type match the other variables inside the class)
- Changed id_to_find from int to uint32_t in Team.cpp

+++ Projects/CoX/Servers/MapServer/NetCommandManager.cpp
@@ -124,7 +124,7 @@ void NetCommandManager::serializeto(BitStream &tgt, const vNetCommand &commands,
     }
     else
     {
-        for(uint32_t i=0; i<(uint32_t)commands.size(); i++)
+        for(uint32_t i=0; i<commands.size(); i++)
         {
             tgt.StorePackedBits(1,i+1);
             tgt.StoreString(commands[i]->m_name);
